refactor(i18n): move language preference functions to languageApi module
